巡查结果的新需求界面,接口完善,逻辑完善ok;智慧消防阻塞性问题调试解决;

+++ OpticalValleyUnite/Classes/UI/PatrolResult/ViewController/YQResultDetailViewController.swift
@@ -13,7 +13,7 @@ import SVProgressHUD
 class YQResultDetailViewController: UIViewController {
     
     /// 属性列表
-    var insResultId : Int = 0
+    var insResultId : Int64 = 0
     
     @IBOutlet weak var pointName: UILabel!
